decouple pipeline and interperteer
