+++ src/query/service/src/pipelines/common/stage.rs
