+++ src/query/service/src/pipelines/pipeline_builder.rs
@@ -91,9 +91,9 @@ use super::processors::transforms::WindowFunctionInfo;
 use super::processors::TransformExpandGroupingSets;
 use crate::api::DefaultExchangeInjector;
 use crate::api::ExchangeInjector;
-use crate::interpreters::copy_commit_data_and_set_finish;
-use crate::interpreters::PlanParam;
+use crate::pipelines::common::copy_append_data_and_set_finish;
 use crate::pipelines::common::fill_missing_columns;
+use crate::pipelines::common::CopyPlanParam;
 use crate::pipelines::processors::transforms::build_partition_bucket;
 use crate::pipelines::processors::transforms::AggregateInjector;
 use crate::pipelines::processors::transforms::FinalSingleStateAggregator;
@@ -225,11 +225,11 @@ impl PipelineBuilder {
         let start = Instant::now();
         stage_table.read_data(table_ctx, &distributed_plan.source, &mut self.main_pipeline)?;
         // append data
-        copy_commit_data_and_set_finish(
+        copy_append_data_and_set_finish(
             ctx,
             &mut self.main_pipeline,
             distributed_plan.required_source_schema.clone(),
-            PlanParam::DistributedCopyIntoTable(distributed_plan.clone()),
+            CopyPlanParam::DistributedCopyIntoTable(distributed_plan.clone()),
             to_table,
             distributed_plan.files.clone(),
             start,
